Code reference

Agent types

Initialisation function

Functions to run an entire simulation

Firms actions

Workers actions

Bank actions

Government actions

Code reference

In this page we document the functions which constitute the bulk of ABCredit.jl functionality.

Agent types

ABCredit.Aggregates — Type

mutable struct Aggregates

A mutable struct representing the aggregates in the model.

Fields:

consumption::Float64: consumption in nominal price

con::Float64: consumption in quantity

init_price::Float64: initial price

price::Float64: price

init_price_k::Float64: initial price of capital goods

price_k::Float64: average price of capital goods

Un::Float64: unemployment

dividends::Float64: dividends

credit_mismatch::Float64: credit mismatch

barK::Float64: barK

defaults::Float64: defaults

defaults_k::Float64: defaults of capital

inflationRate::Float64: inflation rate

Y_nominal_tot::Float64: total nominal output

Y_real::Float64: total real output

gdp_deflator::Float64: GDP deflator

bankruptcy_rate::Float64: bankruptcy rate

totK::Float64: total capital

timestep::Int64: time step

wb::Float64: wage bill

ABCredit.Bank — Type

mutable struct Bank <: AbstractBank

A mutable struct representing a bank.

Fields:

E::Float64: equity

E_threshold::Float64: equity threshold

loans::Float64: loans

profitsB::Float64: profits

dividendsB::Float64: dividends

reserves::Float64: reserves

deposits::Float64: deposits

ABCredit.FirmCapital — Type

mutable struct FirmCapital <: AbstractCapitalFirm

A mutable struct representing a firm's capital.

Fields:

firm_id::Int64: firm id, needed for the job market

Leff_k::Float64: current employees

Y_k::Float64: production

Y_prev_k::Float64: past production

Y_kd::Float64: target quantity demanded

P_k::Float64: prices

A_k::Float64: firm equity

liquidity_k::Float64: firm liquid resources

De_k::Float64: expected demand

deb_k::Float64: firm debts

Q_k::Float64: actual sales

Ftot_k::Float64: total borrowings

interest_r_k::Float64: interest rate on loans

interests_k::Float64: interests paid

Ld_k::Int64: demanded labour

B_k::Float64: financial gap

lev_k::Float64: leverage

vacancies_k::Int64: number of vacancies

inventory_k::Float64: stock of goods

wages_k::Float64: wages paid

stock_k::Float64: stock of goods

income::Float64: the income provided to the owner of the firm

PA::Float64: the PA of the owner of the firm

cons_budget::Float64: consumption budget of the owner of the firm

permanent_income::Float64: time-averaged income

ABCredit.FirmProduction — Type

mutable struct FirmProduction <: AbstractConsumptionFirm

A FirmProduction type.

Fields

firm_id::Int64: firm id, needed for the job market

value_investments::Float64: investments in the firm

investment::Float64: physical capital acquired in the period

K::Float64: physical capital

A::Float64: firm equity

liquidity::Float64: firm liquid resources

capital_value::Float64: capital value

P::Float64: prices

Y_prev::Float64: past production

Yd::Float64: target quantity demanded

Q::Float64: actual sales

Leff::Float64: current employees

De::Float64: expected demand

deb::Float64: firm debts

barK::Float64:

barYK::Float64:

x::Float64:

interest_r::Float64: interest rate on loans

interests::Float64: interests paid

Ftot::Float64: total borrowings

K_dem::Float64: demanded capital

K_des::Float64: desired capital

Ld::Int64: demanded labour

B::Float64: financial gap

lev::Float64: leverage

vacancies::Int64: number of vacancies

Y::Float64: production

wages::Float64: wages paid

stock::Float64: stock of goods

income::Float64: the income provided to the owner of the firm

PA::Float64: the PA of the owner of the firm

cons_budget::Float64: consumption budget of the owner of the firm

permanent_income::Float64: time-averaged income

ABCredit.Government — Type

mutable struct Government <: AbstractGovernment

The Government struct represents a government entity in the model.

Fields:

G::Float64: government spending in subsidies

TA::Float64: total taxes

GB::Float64: government balance

EXP::Float64: expenditure towards firms

bonds::Float64: bonds

bond_interest_rate::Float64: bond interest rate

stock_bonds::Float64: stock of bonds

deficitGDP::Float64: deficit

subsidy::Float64: subsidy

ABCredit.Model — Type

struct Model <: AbstractModel

The Model struct represents the entire model.

Fields:

params::Dict{Symbol, Any}: A dictionary that stores various parameters used in the model.

workers::Vector{Worker}: A vector of Worker objects representing the workers in the model.

consumption_firms::Vector{FirmProduction}: A vector of FirmProduction objects representing the consumption firms in the model.

capital_firms::Vector{FirmCapital}: A vector of FirmCapital objects representing the capital firms in the model.

bank::Bank: An object representing the bank in the model.

gov::Government: An object representing the government in the model.

agg::Aggregates: An object representing the aggregate variables in the model.

ABCredit.Worker — Type

mutable struct Worker <: AbstractWorker

A Worker type.

Fields

w::Float64: earned wages

PA::Float64: household personal assets (saving stock)

Oc::Int64: employment status: Oc(i)=j –> worker j employed by firm i; if i=0, j is unemployed

income::Float64: generic income, it can be wages or subsidy

cons_budget::Float64: consumption budget

permanent_income::Float64: time-averaged income

Initialisation function

ABCredit.initialise_model — Method

initialise_model(W, F, N; params = ABCredit.PARAMS_ORIGINAL)

Initializes the model with the given parameters.

Arguments

W: Number of agents in the W dimension.

F: Number of agents in the F dimension.

N: Number of agents in the N dimension.

params: (optional) Dictionary of parameters for the model. Default is ABCredit.PARAMS_ORIGINAL.

Returns

model: The initialized model.

Functions to run an entire simulation

ABCredit.run_n_sims — Method

run_n_sims(model, T, n_sims)

Run multiple independent simulations of a given model in parallel.

Arguments

model: The model to be simulated.

T: The duration of each simulation.

n_sims: The number of simulations to run.

Returns

An array of ABCreditData objects, each representing the data generated from a single simulation.

ABCredit.run_one_sim! — Method

run_one_sim!(model::AbstractModel, T::Int64; seed = nothing)

Simulates the model for a given number of time steps. In essence it performs a for loop over the complete step function for T times and saves results into an object called 'd'.

Arguments

model::AbstractModel: The model to be simulated.

T::Int64: The number of time steps to simulate.

seed = nothing: Optional seed for the random number generator.

burn_in = 0: Number of time steps to run before starting to collect data.

Returns

d::ABCreditData: The data collector object containing the simulation results.

ABCredit.one_model_step! — Method

one_model_step!(model::AbstractModel)

A complete step of the model simulation.

Arguments

model::AbstractModel: The model object representing the simulation.

Firms actions

ABCredit.firm_produces! — Method

firm_produces!(firm::AbstractCapitalFirm, model::AbstractModel)

This function updates the production of a capital firm based on available resources and model parameters.

Arguments

firm::AbstractCapitalFirm: The consumption firm for which production is being updated.

model::AbstractModel: The model containing the parameters and aggregate variables.

ABCredit.firm_produces! — Method

firm_produces!(firm::AbstractConsumptionFirm, model::AbstractModel)

This function updates the production of a consumption firm based on available resources and model parameters.

Arguments

firm::AbstractConsumptionFirm: The consumption firm for which production is being updated.

model::AbstractModel: The model containing the parameters and aggregate variables.

ABCredit.firms_produce! — Method

firms_produce!(firms, model)

Iterates over a collection of firms and calls the firm_produces! function for each firm.

Arguments

firms: A vector of consumption or capital firms.

model: An abstract model representing the economic environment.

ABCredit.firm_accounting! — Method

firm_accounting!(firm::AbstractConsumptionFirm, model::AbstractModel)

Perform accounting operations for a consumption firm in a given model.

Arguments

firm::AbstractConsumptionFirm: The firm for which accounting operations are performed.

model::AbstractModel: The model in which the firm operates.

Description

This function updates the firm's liquidity, loans, profits, income, dividends, and assets based on the given model parameters.

ABCredit.firm_accounting! — Method

firm_accounting!(firm::AbstractCapitalFirm, model::AbstractModel)

Perform accounting operations for a capital firm in a given model.

Arguments

firm::AbstractCapitalFirm: The firm for which accounting operations are performed.

model::AbstractModel: The model in which the firm operates.

Description

This function updates the firm's liquidity, loans, profits, income, dividends, and assets based on the given model parameters.

Workers actions

ABCredit.worker_gets_paid! — Method

worker_gets_paid!(worker::AbstractWorker, model::AbstractModel)

This function calculates the income of a worker based on their employment status and the model parameters. If the worker is employed, their income is determined by their wage after taxes. If the worker is unemployed, they receive a subsidy. The function also updates the net worth of the worker.

Arguments

worker::AbstractWorker: The worker for whom the income is being calculated.

model::AbstractModel: The model containing the parameters and government data.

ABCredit.workers_get_paid! — Method

workers_get_paid!(workers::Vector{<:AbstractWorker}, model::AbstractModel)

Pay the workers in the given vector by calling worker_gets_paid! for each worker.

Arguments

workers::Vector{<:AbstractWorker}: A vector of workers to be paid.

model::AbstractModel: The model used to calculate the payment.

Bank actions

ABCredit.bank_accounting! — Function

bank_accounting!(bank::AbstractBank, model::AbstractModel)

Update the accounting records of a bank based on the current state of the economy.

Arguments

bank::AbstractBank: The bank object whose accounting records need to be updated.

model::AbstractModel: The model object representing the current state of the economy.

Description

This function updates the profits, dividends, and equity of a bank based on the current state of the economy. It calculates the profits from government loans, pays dividends to shareholders, and distributes the dividends to the income and assets of the firms owners.

Government actions

ABCredit.gov_accounting! — Method

gov_accounting!(gov::AbstractGovernment, model::AbstractModel)

Update the government's accounting by adjusting the government balance, stock of bonds, and deficit to GDP ratio.

Arguments

gov::AbstractGovernment: The government object.

model::AbstractModel: The model object.

ABCredit.gov_adjusts_subsidy! — Method

gov_adjusts_subsidy!(gov::AbstractGovernment, model::AbstractModel)

Adjusts the subsidy amount based on the predicted deficit and the target deficit.

Arguments

gov::AbstractGovernment: The government object.

model::AbstractModel: The model object.

ABCredit.set_bond_interest_rate! — Method

The interest rate on government bonds is set